feat(ir, wasm): add tribute_rt dialect for boxing/unboxing primitives (#187)
* feat(ir): add tribute_rt dialect with 31-bit int/nat types

Add new tribute_rt dialect for backend-independent runtime types:
- Types: int, nat, float, bool, intref, any
- Operations: box_int, unbox_int

Migrate int/nat from tribute to tribute_rt dialect and change from
64-bit to 31-bit representation to align with i31ref for direct boxing.

This prepares for cleaner boxing/unboxing logic by separating runtime
type concerns into a dedicated dialect.



* feat(wasm): add tribute_rt_to_wasm lowering pass

Lower tribute_rt boxing operations to wasm dialect:
- tribute_rt.box_int -> wasm.ref_i31 (i32 -> i31ref)
- tribute_rt.unbox_int -> wasm.i31_get_s (i31ref -> i32)



* feat(wasm): integrate tribute_rt operations into cont_to_wasm pass

Refactor boxing/unboxing operations to use tribute_rt dialect:
- Updated unbox_value_if_needed to use tribute_rt.unbox_int for both Int and Nat types
- Updated box_value_if_needed to use tribute_rt.box_int for Int/Nat primitives
- Simplified unboxing logic by consolidating Int/Nat cases
- Added tribute_rt_to_wasm pass to lowering pipeline to convert tribute_rt ops to wasm

This change follows the multi-phase approach:
- Phase 4: cont_to_wasm generates tribute_rt operations (box_int, unbox_int)
- Phase 5: tribute_rt_to_wasm lowers them to final wasm instructions



* fix(wasm): resolve type consistency between cont_to_wasm and tribute_rt_to_wasm passes

Ensures struct operations use properly lowered types for consistency:
- Added lower_tribute_rt_type helper to convert tribute_rt types to wasm equivalents
- Use lowered types in struct_get operations for resume functions
- Fix box_value_if_needed to handle core.i32 (lowered tribute_rt.int)
- Fix unbox_value_if_needed to use wasm.i31ref instead of tribute_rt.intref
- Add debug logging for type tracking in struct field operations

This resolves type mismatches where struct_new uses lowered core.i32 values
but struct_get expected tribute_rt.int types.


* docs(wasm): document Int/Nat-only restriction for continuation boxing

Add documentation and explicit panic handling for unsupported float/bool
boxing/unboxing in continuation handling. Explains why float needs heap
allocation (no f64ref in WasmGC) and bool could use i31ref.

* refactor(wasm): improve type helper usage and future-proof lowering

Replace string-based type comparisons with tribute_rt type helpers (is_int, is_nat, is_bool, is_float) in gc_types.rs for better maintainability and consistency.

Extend lower_tribute_rt_type to handle bool and float types in cont_to_wasm.rs for future-proofing against additional tribute_rt types.

Add explicit i32 handling for current 31-bit int/nat representation in emit.rs, with i64 path documented as legacy for compatibility.



* chore: change default compilation target to wasm and format conditionals

- Change default target from 'none' to 'wasm' in cli.rs

- Format single-line conditionals in gc_types.rs and cont_to_wasm.rs

---------

+++ tests/e2e_add.rs
@@ -38,9 +38,9 @@ fn test_add_compiles_and_runs() {
             wasmtime::Instance::new(&mut store, &module, &[]).expect("Failed to instantiate");
 
         // Call main directly (add.trb doesn't use print_line, so no _start is generated)
-        // Note: Tribute's Int type compiles to i64 in WASM
+        // Note: Tribute's Int type compiles to i32 in WASM (31-bit signed)
         let main = instance
-            .get_typed_func::<(), i64>(&mut store, "main")
+            .get_typed_func::<(), i32>(&mut store, "main")
             .expect("main function not found");
 
         let result = main.call(&mut store, ()).expect("Execution failed");
@@ -76,9 +76,9 @@ fn main() -> Int { identity(42) }
         let instance =
             wasmtime::Instance::new(&mut store, &module, &[]).expect("Failed to instantiate");
 
-        // Int compiles to i64 in WASM
+        // Int compiles to i32 in WASM (31-bit signed)
         let main = instance
-            .get_typed_func::<(), i64>(&mut store, "main")
+            .get_typed_func::<(), i32>(&mut store, "main")
             .expect("main function not found");
 
         let result = main.call(&mut store, ()).expect("Execution failed");
@@ -118,7 +118,7 @@ fn main() -> Int {
             wasmtime::Instance::new(&mut store, &module, &[]).expect("Failed to instantiate");
 
         let main = instance
-            .get_typed_func::<(), i64>(&mut store, "main")
+            .get_typed_func::<(), i32>(&mut store, "main")
             .expect("main function not found");
 
         let result = main.call(&mut store, ()).expect("Execution failed");
@@ -167,7 +167,7 @@ fn main() -> Int {
             wasmtime::Instance::new(&mut store, &module, &[]).expect("Failed to instantiate");
 
         let main = instance
-            .get_typed_func::<(), i64>(&mut store, "main")
+            .get_typed_func::<(), i32>(&mut store, "main")
             .expect("main function not found");
 
         let result = main.call(&mut store, ()).expect("Execution failed");
@@ -261,7 +261,7 @@ fn main() -> Int {
             wasmtime::Instance::new(&mut store, &module, &[]).expect("Failed to instantiate");
 
         let main = instance
-            .get_typed_func::<(), i64>(&mut store, "main")
+            .get_typed_func::<(), i32>(&mut store, "main")
             .expect("main function not found");
 
         let result = main.call(&mut store, ()).expect("Execution failed");
@@ -311,7 +311,7 @@ fn main() -> Int {
             wasmtime::Instance::new(&mut store, &module, &[]).expect("Failed to instantiate");
 
         let main = instance
-            .get_typed_func::<(), i64>(&mut store, "main")
+            .get_typed_func::<(), i32>(&mut store, "main")
             .expect("main function not found");
 
         let result = main.call(&mut store, ()).expect("Execution failed");
@@ -359,7 +359,7 @@ fn main() -> Int {
             wasmtime::Instance::new(&mut store, &module, &[]).expect("Failed to instantiate");
 
         let main = instance
-            .get_typed_func::<(), i64>(&mut store, "main")
+            .get_typed_func::<(), i32>(&mut store, "main")
             .expect("main function not found");
 
         let result = main.call(&mut store, ()).expect("Execution failed");
@@ -406,7 +406,7 @@ fn main() -> Int {
             wasmtime::Instance::new(&mut store, &module, &[]).expect("Failed to instantiate");
 
         let main = instance
-            .get_typed_func::<(), i64>(&mut store, "main")
+            .get_typed_func::<(), i32>(&mut store, "main")
             .expect("main function not found");
 
         let result = main.call(&mut store, ()).expect("Execution failed");
@@ -633,7 +633,7 @@ fn test_calc_eval() {
             wasmtime::Instance::new(&mut store, &module, &[]).expect("Failed to instantiate");
 
         let main = instance
-            .get_typed_func::<(), i64>(&mut store, "main")
+            .get_typed_func::<(), i32>(&mut store, "main")
             .expect("main function not found");
 
         let result = main.call(&mut store, ()).expect("Execution failed");
